% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoFA.R
\name{CoFA}
\alias{CoFA}
\title{Common Factor Analysis between Functional and Multivariate Data by Adaptive Nuclear Norm Penalty}
\usage{
CoFA(
  y_sample,
  z_sample,
  t_vec,
  y_center = TRUE,
  z_standard = TRUE,
  num_knot,
  num_K1 = NULL,
  num_K2 = NULL,
  gamma = 2,
  BEMA_alpha = 0.2,
  BEMA_beta = 0.1,
  control = list(num_rep = 100, num_fold = 5, num_tau = 100)
)
}
\arguments{
\item{y_sample}{The functional data matrix}

\item{z_sample}{The multivariate data matrix}

\item{t_vec}{The observation points of functional data}

\item{y_center}{A logical value to indicate whether y_sample needs to be be centered (TRUE) or not (FALSE)}

\item{z_standard}{A logical value to indicate whether z_sample needs to be standardized (TRUE) or not (FALSE)}

\item{num_knot}{An integer to choose the total numbers of knots}

\item{gamma}{The parameter used for weights in adaptive nuclear norm penalty}

\item{BEMA_alpha}{One of the two tuning parameters in the BEMA method. It determines a percentage of nonzero eigenvalue in the middle range. Default is 0.2}

\item{BEMA_beta}{One of the two tuning parameters in the BEMA method. It controls the probability of over-estimating K1 and K2. Recommended default is 0.1}

\item{control}{A list that contains the cross validation setting}

\item{user_K1}{An integer for K1 fixed by users. If not specified (NULL), estimation by BEMA method applies}

\item{user_K2}{An integer for K2 fixed by users. If not specified (NULL), estimation by BEMA method applies}
}
\value{
A list that contains all estimation and prediction results
}
\description{
Common Factor Analysis between Functional and Multivariate Data by Adaptive Nuclear Norm Penalty
}
