% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofaForest.R
\name{cofaForest}
\alias{cofaForest}
\title{Fits a Random Forest and calculated CoFA statistics for all pairs of categories.}
\usage{
cofaForest(ntree = 100, cvar, yvar, xvars, data, control.cp = 0.001,
  seed = 4, indices = NULL)
}
\arguments{
\item{ntree}{integer number of trees to fit within the random forest}

\item{cvar}{name of the}

\item{yvar}{name of the column containing a binary outcome variable}

\item{xvars}{vector of names of columns to use as predictors}

\item{data}{a data.frame including columns \code{yvar}, \code{cvar} and \code{xvars}}

\item{indices}{a list of vectors containing the indices of rows to be used when fitting each tree. If \code{NULL}, rows will be randomly sampled with replacement.}
}
\value{
A list of outputs including a matrix of CoFA statistics (\code{freqMat}), a matrix with the total number of times each pair of categories are used for splitting (\code{totalMat}), a list of rpart tree objects (\code{tree}) and a list of the indices used to fit each tree (\code{indices}).
}
\description{
Fits a Random Forest and calculated CoFA statistics for all pairs of categories.
}
