% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofaTest.R
\name{cofaTest}
\alias{cofaTest}
\title{Calculates CoFA statistics for a random forest on the provided data and simulated the null hypothesis distribution by calculating CoFA statistics on data with the categorical variable shuffled}
\usage{
cofaTest(k = 10, data, xvars, yvar, cvar, ntree = 100, seed = NULL)
}
\arguments{
\item{k}{number of iterations to simulate null hypothesis CoFA statistics}

\item{data}{data.frame containing columns \code{xvars}, \code{yvar}, \code{cvar}}

\item{yvar}{name of column containing the outcome variable}

\item{cvar}{name of column containing the categorical variable of interest}

\item{ntree}{number of trees per random forest}

\item{seed}{a random seed}

\item{xvar}{vector of other columns to use in fitting random forests}
}
\description{
Calculates CoFA statistics for a random forest on the provided data and simulated the null hypothesis distribution by calculating CoFA statistics on data with the categorical variable shuffled
}
