% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_application.R
\name{gam_application}
\alias{gam_application}
\title{A function to fit the generalized additive models to compare different types of predicted components
This function loads the saved RData file generated by data_application function}
\usage{
gam_application(which_ear = "right", model_num = 1:9)
}
\arguments{
\item{which_ear}{Either "right" or "left" to choose which ear of data is loaded}

\item{model_num}{A numeric vector of the model numbers which can be positive integers from 1 to 9}
}
\value{
A list of 9 GAM model fits and the adjusted R-square table whose number of rows depend on the num_knot of the saved real_data_app_right, or left, RData file
}
\description{
A function to fit the generalized additive models to compare different types of predicted components
This function loads the saved RData file generated by data_application function
}
