% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofaForest_helpers.R
\name{getAllFreqofLevel}
\alias{getAllFreqofLevel}
\title{Counts the appearances of all categories in a single tree}
\usage{
getAllFreqofLevel(mat)
}
\arguments{
\item{mat}{a matrix returned by \code{getMatix()}}
}
\value{
An integer vector containing the number of splits involving each category
}
\description{
Counts the appearances of all categories in a single tree
}
