% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofaForest_helpers.R
\name{getCofa}
\alias{getCofa}
\title{Calculates the cofa statistic for a pair of levels}
\usage{
getCofa(i, j, mat)
}
\arguments{
\item{i}{an integer}

\item{j}{an integer}

\item{mat}{a matrix of numeric values}
}
\value{
Calculated cofa statistic for level \code{i} and level \code{j} of the categorical variable of interest using the tree split matrix \code{mat}. The cofa statistic is the probability that levels \code{i} and \level{j} split together at a node where the categorical variable of interest is used to split.
}
\description{
This helper function is used by \code{makeCofaMat()}
}
