% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofaForest_helpers.R
\name{getCounts}
\alias{getCounts}
\title{Counts the number of times two categories split together versus split apart necessary to calculate the cofa statistic for a pair of levels}
\usage{
getCounts(i, j, mat)
}
\arguments{
\item{i}{the index of a level}

\item{j}{the index of a level}

\item{mat}{a split matrix returned by \code{getMatrix()}}
}
\value{
A list containing two objects:
\code{together} is the number of times level \code{i} and level \code{j} of the categorical variable of interest split in the same direction
\code{apart} is the number of times level \code{i} and level \code{j} split apart.
}
\description{
This helper function is used by \code{makeCofaMat()} and \code{cofaForest()}
}
