% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_application.R
\name{logistic_application}
\alias{logistic_application}
\title{A function to fit the logistic regression models to compare different types of predicted components
This function loads the saved RData file generated by data_application function}
\usage{
logistic_application(which_ear = "right", num_rep = 1000)
}
\arguments{
\item{which_ear}{Either "right" or "left" to choose which ear of data is loaded}

\item{num_rep}{The number of resampling to overcome highly skewed class distribution in data}
}
\value{
A list that contains the table of the averaged AUC values for different knot settings
}
\description{
A function to fit the logistic regression models to compare different types of predicted components
This function loads the saved RData file generated by data_application function
}
