% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofaForest_helpers.R
\name{makeCofaMat}
\alias{makeCofaMat}
\title{Create a matrix of cofa statistics from the split matric of a single tree}
\usage{
makeCofaMat(mat)
}
\arguments{
\item{mat}{a matrix or vector.}
}
\value{
A matrix where cell (i,j) contains the cofa statistic for level i and level j of the categorical variable of interest. The cofa statistic is the probability that levels i and j split together at a node where the categorical variable of interest is used to split.
}
\description{
Create a matrix of cofa statistics from the split matric of a single tree
}
