% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_omega_est.R
\name{var_omega_est}
\alias{var_omega_est}
\title{A function to estimate the variance of random noise in multivariate data}
\usage{
var_omega_est(z_data_cent, V0_mat, est_beta_vec, est_eta_vec, est_K0)
}
\arguments{
\item{z_data_cent}{The matrix of centered multivariate data}

\item{V0_mat}{The matrix of estimated eigen vectors for common components}

\item{est_beta_vec}{The vector of the estimated beta values}

\item{est_eta_vec}{The vector of the estimated eta values}

\item{est_K0}{The estimated K0 value}
}
\value{
An object of a solution from CVXR R package
}
\description{
A function to estimate the variance of random noise in multivariate data
}
